+++ srbuild/cli/cli.py
@@ -10,12 +10,13 @@ import argparse
 import sys
 import os
 
-# Set up the top-level parser. This needs to happen at import time so that logging verbosity is set early.
-parser = argparse.ArgumentParser(description="Builds this project")
-parser.add_argument("-v", "--verbose", help="Enable verbose logging output", action="store_true")
-parser.add_argument("-vv", "--very-verbose", help="Enable very verbose logging output", action="store_true")
+# Set up a special parser just to intercept -v/-vv. This will not intercept -h, so we duplicate the options
+# in the cli() parser.
+verbosity_parser = argparse.ArgumentParser(add_help=False)
+verbosity_parser.add_argument("-v", "--verbose", action="store_true")
+verbosity_parser.add_argument("-vv", "--very-verbose", action="store_true")
 
-args, _ = parser.parse_known_args()
+args, _ = verbosity_parser.parse_known_args()
 if args.very_verbose:
     G_LOGGER.verbosity = logger.Verbosity.VERBOSE
 elif args.verbose:
@@ -24,6 +25,7 @@ elif args.verbose:
 # TODO: Docstrings
 # Sets up the the command-line interface for the given project/generator combination.
 # When no profile(s) are specified, default_profile will be used.
+# TODO: Make default profile a list, and [] should correspond to all profiles.
 def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="debug"):
     generator = GeneratorType(project)
 
@@ -38,7 +40,7 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
     # TODO: Subprocess management needs to be centralized somewhere.
     def _check_returncode(result: subprocess.CompletedProcess) -> str:
         terminal_width, _ = os.get_terminal_size(0)
-        output = f"\n\n{' Captured stdout '.center(terminal_width, '=')}\n{result.stdout.decode(sys.stdout.encoding)}\n\n{' Captured stderr '.center(terminal_width, '=')}\n{result.stderr.decode(sys.stdout.encoding)}"
+        output = f"\n{' Captured stdout '.center(terminal_width, '=')}\n{result.stdout.decode(sys.stdout.encoding)}\n{' Captured stderr '.center(terminal_width, '=')}\n{result.stderr.decode(sys.stdout.encoding)}"
         if result.returncode:
             G_LOGGER.critical(f"Build failed with:{output}")
         return output
@@ -59,18 +61,22 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
                 G_LOGGER.warning(f"Target: {tgt_name} refers to both an executable and a library. Selecting both.")
         return targets
 
+    # Given argparse's args struct, parses out profile flags, and returns a list of profile names included.
+    def _select_profile_names(args) -> List[str]:
+        return [prof_name for prof_name in project.profiles.keys() if getattr(args, prof_name)]
+
     def configure(args):
         G_LOGGER.info(f"Generating configuration files in build directory: {project.files.build_dir}")
         generator.generate()
 
     @needs_configure
     def build(args):
-        # TODO(1): Build only for the profiles specified.
-        # By default, build all targets
+        # By default, build all targets for the default profile..
         targets = _select_targets(args.targets) or (list(project.libraries.values()) + list(project.executables.values()))
-        G_LOGGER.info(f"Building targets: {[target.name + (' (lib)' if target.is_lib else ' (exe)') for target in targets]}")
+        prof_names = _select_profile_names(args) or [default_profile]
+        G_LOGGER.info(f"Building targets: {[target.name + (' (lib)' if target.is_lib else ' (exe)') for target in targets]} for profiles: {prof_names}")
         G_LOGGER.debug(f"Targets: {targets}")
-        _check_returncode(generator.build(targets))
+        _check_returncode(generator.build(targets, profiles=prof_names))
 
     @needs_configure
     def run(args):
@@ -79,53 +85,81 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
         if args.target not in project.executables:
             G_LOGGER.critical(f"Could not find target: {args.target} in project executables. Note: Available targets are: {list(project.executables.keys())}")
         target = project.executables[args.target]
-        # Build for this profile. TODO: Replace this with correct profile.
-        _check_returncode(generator.build([target], profiles=[default_profile]))
-        G_LOGGER.log(_check_returncode(subprocess.run([target[default_profile].path], capture_output=True)))
+        prof_name = default_profile
+        prof_names = _select_profile_names(args)
+        if prof_names:
+            # Only one profile can be specified to run.
+            assert len(prof_names) == 1
+            prof_name = prof_names[0]
+        _check_returncode(generator.build([target], profiles=[prof_name]))
+        G_LOGGER.info(f"Running target: {args.target}, for profile: {prof_name}:\n{target[prof_name].path}")
+        G_LOGGER.log(_check_returncode(subprocess.run([target[prof_name].path], capture_output=True)))
 
     @needs_configure
     def install(args):
-        # TODO(3): Finish implementation
+        # TODO(1): Finish implementation
         pass
 
     def clean(args):
-        # TODO(3): Finish implementation, per-profile and per-target cleaning.
-        # When supplied no arguments, this will remove the whole build directory for each profile.
-        G_LOGGER.info("Cleaning...")
-        to_remove = [prof.build_dir for prof in project.profiles.values()]
-        # The nuclear option
+        # TODO(3): Finish implementation, add per-target cleaning.
+        to_remove = []
         if args.nuke:
+            # The nuclear option
             to_remove = [project.build_dir]
+            G_LOGGER.info(f"Initiating Nuclear Protocol!")
+        else:
+            # By default, cleans all targets for the default profile.
+            prof_names = _select_profile_names(args) or [default_profile]
+            to_remove = [project.profiles[prof_name].build_dir for prof_name in prof_names]
+            G_LOGGER.info(f"Cleaning targets for profiles: {prof_names}")
+        # Remove
         for path in to_remove:
             G_LOGGER.info(f"\tRemoving {path}")
             project.files.rm(path)
+
+    parser = argparse.ArgumentParser(description="Builds this project")
+    parser.add_argument("-v", "--verbose", help="Enable verbose logging output", action="store_true")
+    parser.add_argument("-vv", "--very-verbose", help="Enable very verbose logging output", action="store_true")
 
     # By setting defaults, each subparser automatically invokes a function to execute it's actions.
     subparsers = parser.add_subparsers()
     # Configure
     configure_parser = subparsers.add_parser("configure", help="Generate build configuration files", description="Generate a build configuration file for the project")
     configure_parser.set_defaults(func=configure)
+
     # Build
     build_parser = subparsers.add_parser("build", help="Build project targets", description="Build one or more project targets")
-    build_parser.add_argument("targets", nargs='*', help="Targets to build. Builds all targets by default", default=[])
+    build_parser.add_argument("targets", nargs='*', help="Targets to build. Builds all targets for the default profile by default", default=[])
+    for prof_name in project.profiles.keys():
+        build_parser.add_argument(f"--{prof_name}", help=f"Build targets using the {prof_name} profile", action="store_true")
     build_parser.set_defaults(func=build)
+
     # Run
-    # TODO: This should accept --profile as arguments. Add each to a mutually exclusive group.
     run_parser = subparsers.add_parser("run", help="Run a project executable", description="Run a project executable")
     run_parser.add_argument("target", help="Target corresponding to an executable")
+    run_profile_group = run_parser.add_mutually_exclusive_group()
+    for prof_name in project.profiles.keys():
+        run_profile_group.add_argument(f"--{prof_name}", help=f"Run the {prof_name} profile's target", action="store_true")
     run_parser.set_defaults(func=run)
+
     # Install
     install_parser = subparsers.add_parser("install", help="Install a project target", description="Install a project target")
     install_parser.set_defaults(func=install)
+
     # Clean
-    clean_parser = subparsers.add_parser("clean", help="Clean project targets", description="Clean one or more project targets. If no arguments are provided, cleans all targets.")
-    clean_parser.add_argument("--nuke", help="The nuclear option. Removes the entire build directory, meaning that the project must be reconfigured before subsequent builds.")
+    clean_parser = subparsers.add_parser("clean", help="Clean project targets", description="Clean one or more project targets. If no arguments are provided, cleans all targets for the default profile.")
+    clean_parser.add_argument("--nuke", help="The nuclear option. Removes the entire build directory, including all targets for all profiles, meaning that the project must be reconfigured before subsequent builds.", action="store_true")
+    for prof_name in project.profiles.keys():
+        clean_parser.add_argument(f"--{prof_name}", help=f"Cleans targets of the {prof_name} profile", action="store_true")
     clean_parser.set_defaults(func=clean)
 
     # Dispatch
     args, unknown = parser.parse_known_args()
+    # If there are unknown arguments, make the parser display an error.
+    # Done this way because a parser option provided after a subparser will
+    # result in an unknown arg that isn't really unknown.
     if unknown:
-        G_LOGGER.critical(f"Unknown arguments: {unknown}")
+        parser.parse_args(unknown)
 
     if hasattr(args, "func"):
         args.func(args)